requirements.txt
fastapi
uvicorn
python-multipart
streamlit
chromadb
groq
pymupdf
python-telegram-bot
python-dotenv
requests
pydantic
pydantic-settings
citeproc-py-styles
citeproc-py
jinja2
sqlalchemy
psycopg2-binary
telethon


